+++ tests/NestedSetTest.php
@@ -33,12 +33,18 @@ class NestedSetTest extends TestCase
 
     public function testConvertToAL()
     {
-        $this->assertTrue(true);
+        $this->assertSame(
+            json_encode($this->al),
+            json_encode($this->converter->toAjacencyList())
+        );
     }
 
     public function testConvertToMP()
     {
-        $this->assertTrue(true);
+        $this->assertSame(
+            json_encode($this->mp),
+            json_encode($this->converter->toMaterializedPath())
+        );
     }
 
     public function testConvertToNS()