+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2022-04-22
+
+### Added
+
+- Nested set to ajacency list convertion
+- Nested set to materialized path convertion
+- Nested set convertion tests
+
 ## [0.5.0] - 2022-04-22
 
 ### Added